+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useMemo } from "react";
 import Box from "@mui/material/Box";
 import Grid from "@mui/material/Grid";
-import AppList from "@/components/AppList";
+import AppList from "@/components/AppGallery";
 import Search from "@/components/SearchBox";
 import Tips from "@/components/Tips";
 import Bookmark from "@/components/Bookmark";